Replace vpc base image from VyOS to generic Linux

See merge request zstackio/zstack-vyos!757

+++ scripts/install-zvr.sh
@@ -11,11 +11,12 @@ err_exit "unable to untar data.tar.gz"
 set -u
 
 ARCH=`uname -m`
-LABLE=''
-[ x"$ARCH" != x"x86_64" ] && LABLE="_$ARCH"
+SUFFIX="_$ARCH"
 
-dpkg -l dnsmasq
-if [ $? -ne 0 -a x"$ARCH" = x"aarch64" ];then
+id -u vyos > /dev/null 2>&1 && USER="vyos" || USER="zstack"
+
+dpkg -l dnsmasq > /dev/null 2>&1
+if [ $? -ne 0 -a x"$ARCH" = x"aarch64" -a x"$USER" = x"vyos" ];then
     rm -rf /etc/dnsmasq.conf
     sudo dpkg -i dnsmasq_pkg/*
 fi
@@ -29,10 +30,10 @@ elif [ "$OSVERSION" = "5.4.80-amd64-vyos" ]; then
   sudo cp  grub.cfg.5.4.80 /boot/grub/grub.cfg
 fi
 
-ZVR_DATA_DIR="/home/vyos/zvr/data"
-ZVR_DATA_DIR_TEMP="/home/vyos/zvr/data_temp"
-if [ -d ${ZVR_DATA_DIR_TEMP} ]; then
-	sudo rm -rf ${ZVR_DATA_DIR_TEMP}
+ZVR_DATA_DIR="/home/$USER/zvr/data"
+ZVR_DATA_DIR_TEMP="/home/$USER/zvr/data_temp"
+if [ -d ${ZVR_DATA_DIR} ]; then
+	sudo rm -rf ${ZVR_DATA_DIR}
 fi
 sudo mkdir -p ${ZVR_DATA_DIR_TEMP}
 tar zxf zvr-data.tar.gz -C ${ZVR_DATA_DIR_TEMP}
@@ -40,11 +41,12 @@ tar zxf zvr-data.tar.gz -C ${ZVR_DATA_DIR_TEMP}
 # rsync data_temp/ to data
 sudo rsync -a ${ZVR_DATA_DIR_TEMP}/ ${ZVR_DATA_DIR}
 if [ -d ${ZVR_DATA_DIR_TEMP} ]; then
-	sudo rm -rf ${ZVR_DATA_DIR_TEMP}
+    sudo rm -rf ${ZVR_DATA_DIR_TEMP}
 fi
+
 if [ -f "${ZVR_DATA_DIR}/hooks/00_exec_hooks.sh" ]; then
-    sudo chmod +x /home/vyos/zvr/data/hooks/00_exec_hooks.sh
-    sudo /bin/bash /home/vyos/zvr/data/hooks/00_exec_hooks.sh
+    sudo chmod +x /home/$USER/zvr/data/hooks/00_exec_hooks.sh
+    sudo /bin/bash /home/$USER/zvr/data/hooks/00_exec_hooks.sh $USER
 fi
 
 exit 0
